File & folder cleanup to match systems folder heirarchy layout.
