+++ plugins/filament_view/core/components/derived/basetransform.h
@@ -16,7 +16,7 @@
 #pragma once
 
 #include <filament/math/quat.h>
-#include "component.h"
+#include "core/components/base/component.h"
 #include "shell/platform/common/client_wrapper/include/flutter/encodable_value.h"
 
 namespace plugin_filament_view {
