+++ plugins/filament_view/core/components/derived/commonrenderable.cc
