+++ plugins/filament_view/core/components/derived/commonrenderable.h
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include "component.h"
+#include "core/components/base/component.h"
 #include "shell/platform/common/client_wrapper/include/flutter/encodable_value.h"
 
 namespace plugin_filament_view {
