+++ plugins/filament_view/core/entity/base/entityobject.cc
