+++ plugins/filament_view/core/entity/derived/model/animation/animation.cc
