+++ plugins/filament_view/core/entity/derived/shapes/cube.cc
