+++ plugins/filament_view/core/entity/derived/shapes/plane.cc
