+++ plugins/filament_view/core/entity/derived/shapes/sphere.h
@@ -18,7 +18,7 @@
 
 #include "shell/platform/common/client_wrapper/include/flutter/encodable_value.h"
 
-#include "core/entity/shapes/baseshape.h"
+#include "core/entity/derived/shapes/baseshape.h"
 
 namespace plugin_filament_view {
