+++ plugins/filament_view/core/systems/derived/collision_system.cc
@@ -15,10 +15,10 @@
  */
 #include "collision_system.h"
 
-#include <core/entity/model/model.h>
-#include <core/entity/shapes/cube.h>
-#include <core/entity/shapes/plane.h>
-#include <core/entity/shapes/sphere.h>
+#include <core/entity/derived/model/model.h>
+#include <core/entity/derived/shapes/cube.h>
+#include <core/entity/derived/shapes/plane.h>
+#include <core/entity/derived/shapes/sphere.h>
 #include <core/systems/ecsystems_manager.h>
 
 #include "filament_system.h"
